`XChainBridge`: Introduce sidechain support (XLS-38): (#4292)
A bridge connects two blockchains: a locking chain and an issuing
chain (also called a mainchain and a sidechain). Both are independent
ledgers, with their own validators and potentially their own custom
transactions. Importantly, there is a way to move assets from the
locking chain to the issuing chain and a way to return those assets from
the issuing chain back to the locking chain: the bridge. This key
operation is called a cross-chain transfer. A cross-chain transfer is
not a single transaction. It happens on two chains, requires multiple
transactions, and involves an additional server type called a "witness".

A bridge does not exchange assets between two ledgers. Instead, it locks
assets on one ledger (the "locking chain") and represents those assets
with wrapped assets on another chain (the "issuing chain"). A good model
to keep in mind is a box with an infinite supply of wrapped assets.
Putting an asset from the locking chain into the box will release a
wrapped asset onto the issuing chain. Putting a wrapped asset from the
issuing chain back into the box will release one of the existing locking
chain assets back onto the locking chain. There is no other way to get
assets into or out of the box. Note that there is no way for the box to
"run out of" wrapped assets - it has an infinite supply.

+++ src/ripple/app/tx/impl/Change.cpp
@@ -577,7 +577,8 @@ Change::activateXahauGenesis()
     }
 
     {
-        ripple::STArray hooks{sfHooks, static_cast<int>(genesis_hooks.size())};
+        ripple::STArray hooks{
+            sfHooks, static_cast<std::size_t>(genesis_hooks.size())};
         int hookCount = 0;
 
         for (auto const& [hookOn, wasmBytes, params] : genesis_hooks)
